refactor: share neural segmenter construction
